Split sync into modules
