+++ src/sync/mod.rs
@@ -1,0 +1,5 @@
+pub mod irq;
+pub mod semaphore;
+
+pub use self::irq::{IrqLock, IrqSpinLock};
+pub use self::semaphore::Semaphore;
